fix security bug
- if no or wrong user is loggedin the ontology manager and his actions are not accessible
- fixes #40

+++ extensions/patapro/PataproController.php
@@ -365,10 +365,15 @@ class PataproController extends OntoWiki_Controller_Component
     {
         $this->view->headLink()->appendStylesheet($this->_componentUrlBase .'css/store.css');
         $this->view->headScript()->appendFile($this->_componentUrlBase .'js/store.js');
-
-        $this->view->url = $this->_config->urlBase;
-
-        $this->view->models = $this->_ontologies;
+        
+        if ($this->_erfurt->getAc()->isActionAllowed('PataproStore')) {
+            $this->view->url = $this->_config->urlBase;
+    
+            $this->view->models = $this->_ontologies;
+        }
+        else {
+            $this->_redirect($this->_config->urlBase . 'index', array());
+        }
     }
 
     public function changemodelAction()
@@ -385,53 +390,57 @@ class PataproController extends OntoWiki_Controller_Component
         $action = urldecode($this->getParam ('do'));
         $backup = urldecode($this->getParam ('backup', false));
 
-        if ("" != $modelName)
-        {
-            $jsonReturnValue['modelName'] = $modelName;
-            $jsonReturnValue['action'] = $action;
-            $jsonReturnValue['modelUri'] = $this->_ontologies[$modelName]['namespace'];
-            $jsonReturnValue['files'] = array();
-            $jsonReturnValue['log'] = array();
-            if ("remove" == $action)
+        if ($this->_erfurt->getAc()->isActionAllowed('PataproStore')) {
+            if ("" != $modelName)
             {
-                if ($this->_store->isModelAvailable($this->_ontologies[$modelName]['namespace']))
+                $jsonReturnValue['modelName'] = $modelName;
+                $jsonReturnValue['action'] = $action;
+                $jsonReturnValue['modelUri'] = $this->_ontologies[$modelName]['namespace'];
+                $jsonReturnValue['files'] = array();
+                $jsonReturnValue['log'] = array();
+                if ("remove" == $action)
                 {
-                    if ('true' == $backup)
+                    if ($this->_store->isModelAvailable($this->_ontologies[$modelName]['namespace']))
                     {
-                        $this->_backupOntology($modelName);
+                        if ('true' == $backup)
+                        {
+                            $this->_backupOntology($modelName);
+                        }
+                        $this->_store->deleteModel($this->_ontologies[$modelName]['namespace']);
+                        $jsonReturnValue['log'][] = "model removed";
+                    } else {
+                        $jsonReturnValue['error'] = "model not available";
                     }
-                    $this->_store->deleteModel($this->_ontologies[$modelName]['namespace']);
-                    $jsonReturnValue['log'][] = "model removed";
-                } else {
-                    $jsonReturnValue['error'] = "model not available";
                 }
-            }
-            if ("add" == $action)
-            {
-                // get ontologies config object
-                $ontologies = $this->_config->ontologies->toArray();
-                $ontologiePath = getcwd() . '/' . $ontologies['folder'] . '/';
-
-                $locator = Erfurt_Syntax_RdfParser::LOCATOR_FILE;
-                $filetype = 'auto';
-                $newType = Erfurt_Store::MODEL_TYPE_OWL;
-
-                // create model
-                $model = $this->_store->getNewModel($this->_ontologies[$modelName]['namespace'], $this->_ontologies[$modelName]['namespace'], $newType);
-                $jsonReturnValue['log'][] = "model added";
-                // connect it with system model
-                $this->_store->addStatement("http://localhost/OntoWiki/Config/", $this->_ontologies[$modelName]['namespace'], "http://ns.ontowiki.net/SysOnt/hiddenImports", array( "value" => "http://ns.ontowiki.net/SysBase/", "type" => "uri"));
-                foreach ($this->_ontologies[$modelName]['files'] as $filename)
+                if ("add" == $action)
                 {
-                    // import data to model
-                    $this->_store->importRdf($this->_ontologies[$modelName]['namespace'], $ontologiePath . $filename, $filetype, $locator);
-                    $jsonReturnValue['files'][] = $filename;
-                    $jsonReturnValue['log'][] = "file " . $filename. " added to model " . $modelName;
+                    // get ontologies config object
+                    $ontologies = $this->_config->ontologies->toArray();
+                    $ontologiePath = getcwd() . '/' . $ontologies['folder'] . '/';
+    
+                    $locator = Erfurt_Syntax_RdfParser::LOCATOR_FILE;
+                    $filetype = 'auto';
+                    $newType = Erfurt_Store::MODEL_TYPE_OWL;
+    
+                    // create model
+                    $model = $this->_store->getNewModel($this->_ontologies[$modelName]['namespace'], $this->_ontologies[$modelName]['namespace'], $newType);
+                    $jsonReturnValue['log'][] = "model added";
+                    // connect it with system model
+                    $this->_store->addStatement("http://localhost/OntoWiki/Config/", $this->_ontologies[$modelName]['namespace'], "http://ns.ontowiki.net/SysOnt/hiddenImports", array( "value" => "http://ns.ontowiki.net/SysBase/", "type" => "uri"));
+                    foreach ($this->_ontologies[$modelName]['files'] as $filename)
+                    {
+                        // import data to model
+                        $this->_store->importRdf($this->_ontologies[$modelName]['namespace'], $ontologiePath . $filename, $filetype, $locator);
+                        $jsonReturnValue['files'][] = $filename;
+                        $jsonReturnValue['log'][] = "file " . $filename. " added to model " . $modelName;
+                    }
                 }
+            } else {
+                $jsonReturnValue['error'] = "no model name";
             }
+        } else {
+            $jsonReturnValue['error'] = "no access rights";
         }
-        else
-            $jsonReturnValue['error'] = "no model name";
 
         echo json_encode($jsonReturnValue);
     }
